Add LDAX instruction and move stack instructions to memory module
