+++ emulator/src/core/emulator/instructions/mod.rs
@@ -2,5 +2,4 @@ mod arithmetic;
 mod logic;
 mod r#move;
 mod controlflow;
-mod stack;
 mod memory;
